gis_project/requirements.txt
asgiref==3.8.1
Django==5.2.3
django-cors-headers==4.7.0
django-leaflet==0.32.0
djangorestframework==3.16.0
numpy==2.3.0
psycopg2-binary==2.9.10
shapely==2.1.1
sqlparse==0.5.3
tzdata==2025.2
